Moved FgsCliTest to 'search' package.
