+++ tetrad-cli/src/test/java/edu/cmu/tetrad/cli/search/FgsCliTest.java
@@ -16,9 +16,8 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.cli;
+package edu.cmu.tetrad.cli.search;
 
-import edu.cmu.tetrad.cli.search.FgsCli;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
